third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-import-block-but-self-nested.html

<!DOCTYPE html>
<html>
<head>
<script>
</script>
<link id="target" rel="import" href="resources/child-having-self-policy.html">
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
shouldBe("document.inlineScriptHasRun", "undefined");
shouldBeTrue("document.externalScriptHasRun");
shouldBe("document.corsExternalScriptHasRun", "undefined");
shouldBe("document.evalFromInlineHasRun", "undefined");
shouldBe("document.evalFromExternalHasRun", "undefined");
shouldBe("document.evalFromCorsExternalHasRun", "undefined");
</script>
</body>
</html>
